refactor(db): unified-database schema initializer 분리 + dead code 정리
# 변경
- data/models/schema-initializer.ts (신규, 92 lines)
  - getSchemaSql(): 002-schema.sql 파일 탐색 + LEGACY_SCHEMA fallback
  - initSchema(pool): 스키마 적용 + agent_usage_logs FK migration +
    pg_trgm GIN 인덱스 — 모두 멱등
- data/models/unified-database.ts
  - initSchema / getSchemaSql 제거 (helper 위임)
  - retryQuery 제거 — 모든 사용처가 repository 위임으로 이행 후 dead code
  - 관련 dead utilities 정리: SpanLike / WithSpanLike / cachedWithSpan /
    getWithSpan / QueryParam (retryQuery 전용)
  - SCHEMA_FILE_RELATIVE_PATH 상수 / fs / path / LEGACY_SCHEMA import 제거
  - QueryResult / withRetry import 제거 (retryQuery 만 사용)

# 결과
- unified-database.ts: 645 → 532 lines (raw, -113), max-lines warning 해소
- 전체 backend lint warnings: 5 → 4 (max-lines only)

# 검증
- tsc 0 / lint 0 errors / 0 warnings (unified-database)
- 전체 회귀: 97 suites / 1604 tests / 0 failed

# 후속 (별도 PR)
- ChatService.ts (518), agent-loop-strategy.ts (498),
  sockets/handler.ts (489), request-handler.ts (472)
